frontend/requirements.txt
Flask==2.0.1
kafka-python==2.0.2
numpy==1.23.5
opencv-python==4.8.1.78
Werkzeug==2.0.2